requirements.txt
aiocache==0.12.2
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
humanize==4.9.0
idna==3.6
iso8601==1.1.0
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
starlette-admin==0.12.2
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
yarl==1.9.4
